style(kotlin): improve codes readability.

+++ codes/kotlin/chapter_greedy/max_capacity.kt
@@ -19,8 +19,8 @@ fun maxCapacity(ht: IntArray): Int {
     // 循环贪心选择，直至两板相遇
     while (i < j) {
         // 更新最大容量
-        val cap = (min(ht[i].toDouble(), ht[j].toDouble()) * (j - i)).toInt()
-        res = max(res.toDouble(), cap.toDouble()).toInt()
+        val cap = min(ht[i], ht[j]) * (j - i)
+        res = max(res, cap)
         // 向内移动短板
         if (ht[i] < ht[j]) {
             i++
